requirements.txt
altgraph==0.17
cffi==1.14.0
cryptography==2.8
macholib==1.14
Pillow==8.1.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
qrcode==6.1
shiboken6==6.0.1
six==1.14.0
